Create Ceiling-in-a-sorted-array.py
Given a sorted array and a value x, the ceiling of x is the smallest element in an array greater than or equal to x, and the floor is the greatest element smaller than or equal to x. Assume that the array is sorted in non-decreasing order. Write efficient functions to find the floor and ceiling of x. 
Examples : 

For example, let the input array be {1, 2, 8, 10, 10, 12, 19}
For x = 0:    floor doesn't exist in array,  ceil  = 1
For x = 1:    floor  = 1,  ceil  = 1
For x = 5:    floor  = 2,  ceil  = 8
For x = 20:   floor  = 19,  ceil doesn't exist in array
In the below methods, we have implemented only ceiling search functions. Floor search can be implemented in the same way.

(Binary Search) 

Instead of using linear search, binary search is used here to find out the index. Binary search reduces the time complexity to O(Logn).