Create Check-reversing-sub-array-make-array-sorted.py
Given an array of n distinct integers. The task is to check whether reversing any one sub-array can make the array sorted or not. If the array is already sorted or can be made sorted by reversing any one subarray, print “Yes“, else print “No“.

Examples: 

Input : arr [] = {1, 2, 5, 4, 3}
Output : Yes
By reversing the subarray {5, 4, 3}, the array will be sorted.

Input : arr [] = { 1, 2, 4, 5, 3 }
Output : No

 Linear time solution (O(n)):

The idea to solve this problem is based on the observation that if we perform one rotation of any subarray in the sorted array (increasing order), then we there will be exactly one subarray which will be in decreasing order. So, we have to find that rotated subarray and perform one rotation on it. Finally check if the array becomes sorted or not.

Initialize two variables x and y with -1.
Iterate over the array.
Find the first number for which a[i] > a[i+1] and store it into x. 
Similarly, Store index i+1 as well into y, As this will keep track of the ending of the subarray which is needed to reverse.
Check if x == -1 then array is already sorted so return true.
Otherwise, reverse the array from index x to index y.
Traverse the array to check for every element is sorted or not.
If not sorted, return false.
Finally, return true.