Create Segregate-even-and-odd-elements-in-a-linked-list.py
https://www.geeksforgeeks.org/segregate-even-and-odd-elements-in-a-linked-list/

Given a Linked List of integers, write a function to modify the linked list such that all even numbers appear before all the odd numbers in the modified linked list. Also, keep the order of even and odd numbers the same.

Examples: 

Input: 17->15->8->12->10->5->4->1->7->6->NULL
Output: 8->12->10->4->6->17->15->5->1->7->NULL
Even nodes in the linked list are 8, 12, 10, 4 and 6
Odd Nodes in the linked list are 17, 15, 1 and 7
In the output list, we have all the even nodes first (in the same
order as input list, then all the odd nodes of the list (in the same
order as input list)

Input: 8->12->10->5->4->1->6->NULL
Output: 8->12->10->4->6->5->1->NULL

// If all numbers are even then do not change the list

Input: 8->12->10->NULL
Output: 8->12->10->NULL

// If all numbers are odd then do not change the list

Input: 1->3->5->7->NULL
Output: 1->3->5->7->NULL

Method 1 
The idea is to get a pointer to the last node of list. And then traverse the list starting from the head node and move the odd valued nodes from their current position to end of the list.